fix: Improved update logic
- Improved update logic

+++ launcher.py
@@ -14,8 +14,11 @@ import subprocess
 import zipfile
 import io
 import shutil
-from Dot_Matrix_Panel.outsourced_functions import read, migrate_config, create_userdata, check_for_updates, create_folders
+from Dot_Matrix_Panel.outsourced_functions import read, migrate_config, create_userdata, check_for_updates, create_folders, check_file_access
 from Dot_Matrix_Panel.logger import logger
+import Dot_Matrix_Panel.global_variables as global_variables
+
+global_variables.project_dir = project_dir
 
 def check_internet_connection(url="https://www.google.com", timeout=5):
     try:
@@ -26,10 +29,12 @@ def check_internet_connection(url="https://www.google.com", timeout=5):
 
 def update():
     print("Updating...")
+    logger.info("Updating...")
     branch = "master"
     repo = "johnnyjack123/dot-matrix-info-display"
     folder_to_extract = "Dot_Matrix_Panel"
-    target_folder = "./Dot_Matrix_Panel"
+    target_folder = "Dot_Matrix_Panel"
+    state = True
 
     # 1. ZIP vom GitHub-Branch herunterladen
     zip_url = f"https://github.com/{repo}/archive/refs/heads/{branch}.zip"
@@ -38,6 +43,7 @@ def update():
 
     if response.status_code != 200:
         print(f"❌ Download error: {response.status_code}")
+        logger.error(f"❌ Download error: {response.status_code}")
         return
 
     # 2. ZIP im Arbeitsspeicher entpacken
@@ -48,30 +54,64 @@ def update():
         # 3. Temporär entpacken
         tmp_dir = "_tmp_update_dir"
         if os.path.exists(tmp_dir):
-            shutil.rmtree(tmp_dir)
+            try:
+                check_file_access(tmp_dir)
+                shutil.rmtree(tmp_dir)
+            except PermissionError  as e:
+                logger.error(f"Permission error: {e}")
+
         os.makedirs(tmp_dir, exist_ok=True)
 
         zip_ref.extractall(tmp_dir)
-
-    # 4. Pfad zum zu extrahierenden Ordner
+        logger.info("Extracted update in tmp dir.")
     source_folder = os.path.join(tmp_dir, extracted_dirname, folder_to_extract)
 
     if not os.path.exists(source_folder):
         print(f"❌ Folder '{folder_to_extract}' not found in ZIP!")
+        logger.error(f"❌ Folder '{folder_to_extract}' not found in ZIP!")
         return
 
-    # 5. Zielordner ggf. löschen
+    old_main_path = os.path.join("tmp", "old_files", "main")
+
     if os.path.exists(target_folder):
-        shutil.rmtree(target_folder)
-
-    # 6. Ordner verschieben
-    shutil.move(source_folder, target_folder)
-    print(f"✅ '{folder_to_extract}' was extracted in '{target_folder}'.")
-
-    # 7. Aufräumen
-    shutil.rmtree(tmp_dir)
-    print("🧹 Remove temporary files.")
+        try:
+            check_file_access(target_folder)
+            check_file_access(old_main_path)
+            shutil.move(target_folder, old_main_path)
+            logger.info("Moved old version in tmp.")
+        except Exception as e:
+            state = False
+            logger.error(f"Error in moving old version of main folder: {e}. Program stays on old version.")
+        if not state:
+            pass
+        else:
+            try:
+                check_file_access(source_folder)
+                check_file_access(target_folder)
+                shutil.move(source_folder, target_folder)
+                logger.info("Moved new version in root directory.")
+            except Exception as e:
+                check_file_access(source_folder)
+                check_file_access(target_folder)
+                shutil.move(old_main_path, target_folder)
+                logger.error(f"Error in moving new version of main folder: {e}. Programm stays on old version.")
+            if not state:
+                pass
+            else:
+                try:
+                    old_main_dir_path = os.path.join("tmp", "old_files", "main")
+                    check_file_access(tmp_dir)
+                    check_file_access(old_main_dir_path)
+                    shutil.rmtree(tmp_dir)
+                    shutil.rmtree(old_main_dir_path)
+                    print("🧹 Remove temporary files.")
+                    logger.info("Removed old version.")
+                except PermissionError as e:
+                    logger.error(f"Permission error: {e}")
+    else:
+        logger.error("Target folder doesn´t exists.")
     launch_app()
+
 
 def launch_app():
     subprocess.Popen(["python", "Dot_Matrix_Panel/Dot-Matrix_Main.py"])
